K-Means Clustering in Python
The notebook looks at a dataset on Energy and Electricity sources for countries in Asia Pacific Region, based on which they are clustered depending on electricity sources and other characteristics.

Many thanks to  Irfan Khalid who gave the guidelines in his article: Introduction to Clustering in Python for Beginners in Data Science.