AfterWork Data Science Fellowship Capstone Project
This is a notebook prepared following CRISP-DM 1.0 methodology to help in the analysis of data for the music tracks hosted on Spotify platform. The purpose of the analysis is to help in classification of music genres as well as predict tracks popularity based on the available historical data.

The project has been submitted in partial fulfilment of certification requirements by AfterWork Data Science Fellowship as well as to showcase some Data Science skills learned during the fellowship.